move ensure_color, ensure_grayscale to public utils
